requirements.txt
numpy
onnxruntime
pillow
torch 
torchvision